+++ src/Nova/EmailTrackingTool.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace HenryAvila\EmailTracking\Nova;
+namespace HenryAvila\LaravelNovaEmailTracking\Nova;
 
-use HenryAvila\EmailTracking\Models\Email;
-use HenryAvila\EmailTracking\Policies\EmailPolicy;
+use HenryAvila\LaravelNovaEmailTracking\Models\Email;
+use HenryAvila\LaravelNovaEmailTracking\Policies\EmailPolicy;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
 
-class EmailTrackingTool extends \Laravel\Nova\Tool
+class LaravelNovaEmailTrackingTool extends \Laravel\Nova\Tool
 {
     public static string $emailResource = EmailResource::class;
     public string $emailPolicy = EmailPolicy::class;
@@ -37,7 +37,7 @@ class EmailTrackingTool extends \Laravel\Nova\Tool
         //
     }
 
-    public function emailResource(string $emailResource): EmailTrackingTool
+    public function emailResource(string $emailResource): LaravelNovaEmailTrackingTool
     {
         static::$emailResource = $emailResource;