+++ src/Traits/ModelWithEmailsSenderTrait.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace HenryAvila\EmailTracking\Traits;
+namespace HenryAvila\LaravelNovaEmailTracking\Traits;
 
-use HenryAvila\EmailTracking\Models\Email;
+use HenryAvila\LaravelNovaEmailTracking\Models\Email;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Relations\MorphMany;